x64: In constructor 'idEventArg::idEventArg(...) loses precision

+++ neo/d3xp/gamesys/Class.cpp
@@ -892,7 +892,7 @@ bool idClass::ProcessEventArgs(const idEventDef *ev, int numargs, ...)
 {
 	idTypeInfo	*c;
 	int			num;
-	int			data[ D_EVENT_MAXARGS ];
+	intptr_t		data[ D_EVENT_MAXARGS ];
 	va_list		args;
 
 	assert(ev);
@@ -1010,7 +1010,7 @@ bool idClass::ProcessEvent(const idEventDef *ev, idEventArg arg1, idEventArg arg
 idClass::ProcessEventArgPtr
 ================
 */
-bool idClass::ProcessEventArgPtr(const idEventDef *ev, int *data)
+bool idClass::ProcessEventArgPtr(const idEventDef *ev, intptr_t *data)
 {
 	idTypeInfo	*c;
 	int			num;
